+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.2.1
+
+- Fixed maximum recursion depth exceeded when calling the original method of a
+  patched class calls `super()`.
+
 ## v0.2.0
 
 - Added compatibility for Indico v3.3.